📦  Added FICO score extraction
We needed to have the configs and then we could run the tests to see
what the data was all about, and then make the extraction functions, and
it all worked out. We still have to get the 'frozen' credit figured out,
but we're working on that with CRS Credit.