⚙️  Pulled `requestid` from Headers
The CRS Credit support guys would like to know what the `reauestId` is
on the calls - so they can track them down, and it's returned in the
headers, so we now extract that, and we return it as a top-level value
in the responses. This will make it so much easier to track.